backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings>=2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
openai==2.8.1
python-docx==1.1.0
python-pptx==0.6.23
pytest==7.4.3
hypothesis==6.92.1
python-dotenv==1.0.0
requests==2.31.0
